Move config parsing to a dedicated pkg
